+++ packages/jazz-browser/CHANGELOG.md
@@ -1,5 +1,22 @@
 # jazz-browser
 
+## 0.6.3
+
+### Patch Changes
+
+- Fix migration changes being lost on loaded account
+- Updated dependencies
+  - cojson@0.6.6
+
+## 0.6.2
+
+### Patch Changes
+
+- Fix loading of accounts
+- Updated dependencies
+  - cojson@0.6.5
+  - jazz-autosub@0.6.1
+
 ## 0.6.1
 
 ### Patch Changes