+++ packages/jazz-react/CHANGELOG.md
@@ -1,5 +1,23 @@
 # jazz-react
 
+## 0.5.4
+
+### Patch Changes
+
+- Fix migration changes being lost on loaded account
+- Updated dependencies
+  - cojson@0.6.6
+  - jazz-browser@0.6.3
+
+## 0.5.3
+
+### Patch Changes
+
+- Fix loading of accounts
+- Updated dependencies
+  - cojson@0.6.5
+  - jazz-browser@0.6.2
+
 ## 0.5.2
 
 ### Patch Changes